add search discussion feature
Add a library for search feature that is meilisearch to make it work.

Since we use watch() to check any changes occur in the input search, it will send a request for every character that we typed in. Therefore, use a lodash.debounce library to delay it so that only send a request for 1s after typing.